% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_renviron.R
\name{create_renviron}
\alias{create_renviron}
\title{Create .Renviron file}
\usage{
create_renviron(path)
}
\arguments{
\item{path}{}
}
\description{
Create .Renviron file
}
\examples{
create_renviron("./")
}
